Use worker type module for cross browser compat
